--- a/Manatee.Json/LinqExtensions.cs
+++ b/Manatee.Json/LinqExtensions.cs
@@ -22,7 +22,15 @@ namespace Manatee.Json
 			json.AddRange(results);
 			return json;
 		}
-
+		/// <summary>
+		/// Converts an <see cref="IDictionary{String, JsonValue}"/> into a <see cref="JsonObject"/>.
+		/// </summary>
+		/// <param name="results">An <see cref="IDictionary{String, JsonValue}"/></param>
+		/// <returns>An equivalent <see cref="JsonObject"/></returns>
+		public static JsonObject ToJson(this IDictionary<string, JsonValue> results)
+		{
+			return new JsonObject(results);
+		}
 		/// <summary>
 		/// Converts an <see cref="IEnumerable{T}"/> of <see cref="KeyValuePair{String, JsonValue}"/> returned from a
 		/// LINQ query back into a <see cref="JsonObject"/>.