--- a/Manatee.Json/Serialization/Internal/AutoRegistration/ArraySerializationDelegateProvider.cs
+++ b/Manatee.Json/Serialization/Internal/AutoRegistration/ArraySerializationDelegateProvider.cs
@@ -18,15 +18,22 @@ namespace Manatee.Json.Serialization.Internal.AutoRegistration
 
 		private static JsonValue _Encode<T>(T[] array, JsonSerializer serializer)
 		{
-			var json = new JsonArray();
-			json.AddRange(array.Select(serializer.Serialize));
-			return json;
+			var values = new JsonValue[array.Length];
+			for (int ii = 0; ii < array.Length; ++ii)
+			{
+				values[ii] = serializer.Serialize(array[ii]);
+			}
+			return new JsonArray(values);
 		}
 		private static T[] _Decode<T>(JsonValue json, JsonSerializer serializer)
 		{
-			var list = new List<T>();
-			list.AddRange(json.Array.Select(serializer.Deserialize<T>));
-			return list.ToArray();
+			var array = json.Array;
+			var values = new T[array.Count];
+			for (int ii = 0; ii < array.Count; ++ii)
+			{
+				values[ii] = serializer.Deserialize<T>(array[ii]);
+			}
+			return values;
 		}
 	}
 }