… Manatee.Json/Serialization/Internal/AutoRegistration/DictionarySerializationDelegateProvider.cs
… Manatee.Json/Serialization/Internal/SerializerCache.cs
… Manatee.Json/Serialization/Internal/SerializerMethodPair.cs
… Manatee.Json/Serialization/Internal/Serializers/AutoSerializer.cs
… Manatee.Json/Serialization/Internal/Serializers/ReferencingSerializer.cs
… Manatee.Json/Serialization/Internal/Serializers/ReflectionCache.cs