Various serialization performance improvements (and some deserialization improvements).

- Remove excess allocations in the `JsonValue.ToString` code path
  - Adds new internal `AppendString` and `AppendIndentedString` methods to use an existing `StringBuilder`
- Refactor `InsertEscapeSequences` to use a `StringBuilder` (if necessary)
- Removed LINQ on hot path in `AbstractionMap`
- Removed LINQ on hot path in `CustomSerializations`
- Removed LINQ from and decreased allocations in auto serializers/deserializers
- `SerializationPairCache` now tracks the object itself. **This should be scrutinized to ensure it does not leak.**
- `SerializerMethodPair` now builds `Func<>`'s instead of using `MemberInfo.Invoke`
- Removed LINQ on hot path in `AutoSerializer`
  - Also fixed a bug (?) where the input `JsonObject` was mutated, switched to a set which tracks visited properties instead. **This should be scrutinized as it seems to change existing behavior, but may be a bug.**
- Removed LINQ on hot path in `ReflectionCache`

Manatee 9.5.1 vs Json.NET 10.X:
```
                               Method |      Mean |     Error |    StdDev |     Gen 0 |    Gen 1 |   Gen 2 | Allocated |
------------------------------------- |----------:|----------:|----------:|----------:|---------:|--------:|----------:|
           JsonNET_JObject_FromObject |  23.67 ms | 0.8877 ms |  2.617 ms |  906.2500 | 375.0000 | 93.7500 |   5.05 MB |
     Manatee_JsonSerializer_Serialize | 241.50 ms | 7.8860 ms | 23.250 ms | 1312.5000 | 250.0000 | 62.5000 |   7.92 MB |
 New_Manatee_JsonSerializer_Serialize |  28.60 ms | 1.3421 ms |  3.957 ms |  687.5000 | 250.0000 |       - |   4.47 MB |
```
Approximate 8.4x speedup, 1.8x reduction in memory allocated along the `Serialize` path.

```
                         Method |     Mean |     Error |    StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
------------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|----------:|
       JsonNET_JObject_ToString | 6.890 ms | 0.1592 ms | 0.4694 ms | 445.0000 | 304.3750 | 163.7500 |    2.9 MB |
 New_Manatee_JsonValue_ToString | 1.100 ms | 0.0209 ms | 0.0529 ms |  46.8884 |  29.0293 |  20.5212 | 293.48 KB |
```
Difference in speed from 9.5.1 is too great to measure (takes ~10s/op), but basically 10 s to 1.1 ms!